+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.0.9] - 2023-04-25
 
 ### Added
 
@@ -125,6 +125,7 @@ which extend from the abstract `Election` class.
 
 - First unstable version of the SDK for testing purposes
 
+[0.0.9]: https://github.com/vocdoni/vocdoni-sdk/releases/tag/v0.0.9
 [0.0.8]: https://github.com/vocdoni/vocdoni-sdk/releases/tag/v0.0.8
 [0.0.7]: https://github.com/vocdoni/vocdoni-sdk/releases/tag/v0.0.7
 [0.0.6]: https://github.com/vocdoni/vocdoni-sdk/releases/tag/v0.0.6